update on registration, department, and vital signs

+++ app/Http/Controllers/DepartmentController.php
@@ -38,6 +38,12 @@ class DepartmentController extends Controller
 
     public function destroy(Department $department)
     {
+        if ($department->users_count > 0) {
+            return response()->json([
+                'message' => 'You cannot delete delete this department.'
+            ]);
+        }
+
         $department->delete();
 
         return response()->json(['message' => 'Department successfully deleted.']);
